[Text] Edits to Character Dialogue
Why this change is needed: The edits and rewrites in this
commit explain that TMs are now reusable, rewriting the
old dialogue to match the new mechanics.

It also adjusts new information about evolution stones
to match the style guide of the Pokemon Crystal dialogue:
Pokemon names, Key Places, and Item names in all caps with
no lower case.

Notes for reviewers: Hi! My name is Joe, I am
a Technical Writer and Editor. I have tried out your
work on Pokemon Crystal 2.0 and am enjoying the
quality of life changes.

With these changes, I have not been able to rebuild
the .maps or other key files with the make commands. It
returns a "Version 6 object file" incompatible error.
Please feel free to delete the change if they will break
the game files.